Extract a modify-cargo-toml library for reuse
